backend/pyproject.toml
# backend/pyproject.toml
[tool.poetry]
name = "bonass-soccer-backend"
version = "1.0.0"
description = "Football Team Management Backend"
authors = ["Bonass Soccer Team"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2"
djangorestframework = "^3.14"
django-cors-headers = "^4.2"
django-filter = "^23.3"
psycopg2-binary = "^2.9"
celery = "^5.3"
redis = "^5.0"
django-celery-beat = "^2.5"
django-celery-results = "^2.5"
drf-spectacular = "^0.27"
djangorestframework-simplejwt = "^5.3"
django-environ = "^0.11"
python-dotenv = "^1.0"
gunicorn = "^21.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4"
pytest-django = "^4.5"
pytest-factoryboy = "^2.5"
factory-boy = "^3.3"
faker = "^21.1"
black = "^23.11"
flake8 = "^6.1"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --tb=short"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
